i. new coalescing rules - use knob new_coalescing to turn it on
earlier,even if uncoalesced accesses from the threads of a warp mapped to the
same cache block (for the same instruction), there would be one request for
each thread.  now, the number of requests is equal to the number of unique
cache blocks that are accessed. earlier behavior was fine when GPUs didn't have
data caches, with data caches, one request for each thread can result in
serialization due to port conflicts. new code can reduce conflicts when several
uncoalesced accesses map to the same cache block. note that new implementation
is not strictly in line with what the cuda computing guide states.

ii. appropriate behavior for writes to global memory - set knob
compute_capability to 2.0

a. when a write to global memory hits in l1, the line is evicted and inserted into
l2.
b. when a write to local/global memory misses in l1, the line is directly written
to l2.
c. send acks for completion of writes to global and local memory (have to
rethink about sending acks for local mem writes), sending acks for global
memory writes is necessary for implementing global fence instructions.





git-svn-id: https://svn.research.cc.gatech.edu/macsim/trunk@357 76505255-db3a-46bb-8524-017587b0c4bf